fix: bill add and udpate

+++ pkg/model/bill.go
@@ -33,8 +33,8 @@ type BillRequestFilter struct {
 }
 
 type AddBillRequest struct {
-	StoreId         int             `json:"store_id" binding:"required"`
-	State           int8            `json:"state"`
+	StoreId         int32           `json:"store_id" binding:"required"`
+	State           int32           `json:"state"`
 	PaymentDueDate  *string         `json:"payment_due_date" `
 	PaymentDate     *string         `json:"payment_date" `
 	Discount        string          `json:"discount" binding:"required"`
